docker: move recoverable error proto to shared structs
