+++ plugins/drivers/client.go
@@ -15,6 +15,7 @@ import (
 	"github.com/hashicorp/nomad/plugins/drivers/proto"
 	"github.com/hashicorp/nomad/plugins/shared"
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
+	sproto "github.com/hashicorp/nomad/plugins/shared/structs/proto"
 	"google.golang.org/grpc/status"
 )
 
@@ -138,7 +139,7 @@ func (d *driverPluginClient) StartTask(c *TaskConfig) (*TaskHandle, *cstructs.Dr
 	if err != nil {
 		st := status.Convert(err)
 		if len(st.Details()) > 0 {
-			if rec, ok := st.Details()[0].(*proto.RecoverableError); ok {
+			if rec, ok := st.Details()[0].(*sproto.RecoverableError); ok {
 				return nil, nil, structs.NewRecoverableError(err, rec.Recoverable)
 			}
 		}
