+++ plugins/drivers/proto/driver.proto
@@ -7,12 +7,6 @@ import "google/protobuf/timestamp.proto";
 
 import "github.com/hashicorp/nomad/plugins/shared/hclspec/hcl_spec.proto";
 
-// RecoverableError is used with a grpc Status to indicate if the error is one
-// which is recoverable and can be reattempted by the client.
-message RecoverableError {
-    bool recoverable = 1;
-}
-
 // Driver service defines RPCs used to communicate with a nomad runtime driver.
 // Some rpcs may not be implemented by the driver based on it's capabilities.
 service Driver {
